eval/requirements.txt
datasets>=2.19.0
httpx>=0.27.0
pandas>=2.0.0
Pillow>=10.0.0
tqdm>=4.66.0
click>=8.1.0
python-dotenv>=1.0.0
